requirements.txt
deepl
python-dotenv
questionary
